% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflect-package.R
\docType{data}
\name{egfr_data}
\alias{egfr_data}
\title{Dataset of a EGFR-mutant Cohort}
\format{
The format is a list containing the following elements:
\itemize{
\item mat_value: (Phospho-)protein expression Z scores from RPPA mesurement (260 samples, 179 features).
\item df_sample: A data frame for sample annotations, having 4 columns: SampleID, TumorType, Stratification, Category.
\item df_feature: A data frame for feature annotations, having 3 columns: Feature, FunctionScore, IsActionable.
}
}
\description{
This dataset is (phospho-)protein expression Z scores of an EGFR-mutant cohort measured by RPPA.
}
\examples{
egfr_data

}
\references{
Li X., et al. (2020) Precision combination therapies from recurrent oncogenic co-alterations \doi{https://doi.org/10.1101/2020.06.03.132514}.
}
\keyword{datasets}
