% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflect-package.R
\docType{data}
\name{egfr_result}
\alias{egfr_result}
\title{Result of a EGFR-mutant Cohort from REFLECT}
\format{
The format is a list containing the following elements:
\itemize{
\item gapstat_bestwbound: Gap statistic and best wbound.
\item shc: An object of class hclust.
\item mat_recur_pval: A matrix of recurrence P values.
\item mat_recur_pval: Rcurrent and acitionable features.
\item df_coaltered_targets: Co-altered, recurrent, actionable targets.
}
}
\description{
This is the result from running reflect_pipeline on the egfr_data dataset.
}
\examples{
egfr_result

}
\references{
Li X., et al. (2020) Precision combination therapies from recurrent oncogenic co-alterations \doi{https://doi.org/10.1101/2020.06.03.132514}.
}
\keyword{datasets}
