% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_best_wbound}
\alias{get_best_wbound}
\title{Get the Best Tunning Parameter}
\usage{
get_best_wbound(
  mat_value,
  dissimilarity = c("squared.distance", "absolute.value"),
  wbounds = NULL,
  nperms = 10,
  min_number_features = 10
)
}
\arguments{
\item{mat_value}{A matrix of expression/alteration with samples as rows and features as columns.}

\item{dissimilarity}{A string for the type of dissimilarity, either "squared.distance" or "absolute.value". Default "squared.distance".}

\item{wbounds}{The sequence of tuning parameters to consider. If NULL, then a default sequence seq(1.1, sqrt(ncol(mat_value)), 100) will be used. If non-null, should be greater than 1.}

\item{nperms}{The number of permutations to perform. Default 10.}

\item{min_number_features}{The minimal number of features that the best wbound could generate. Only wbounds that generates more than this number of features can considered. Default 10.}
}
\value{
\item{df_gapstat}{A data frame that contains columns: Wbound, NumberFeatures, GapStat, SD.} \item{best_wbound}{The best tunning parameter whose gap statistic is within one-standard-error to the left of the maximum gap statistic.}
}
\description{
This employs a permutation approach to select the tuning parameter, which controls sparsity of features through L1 regularization. Note this calculation may take a long time.
}
\examples{
library(reflect)
mat_value <- egfr_data$mat_value
gapstat_bestwbound <- get_best_wbound(mat_value)
gapstat_bestwbound$best_wbound
}
