% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_coaltered_targets}
\alias{get_coaltered_targets}
\title{Get Co-altered Targets from Stratification Status and Recurrent/Actionable Features}
\usage{
get_coaltered_targets(df_sample, df_recur_actionable)
}
\arguments{
\item{df_sample}{A data frame of samples including tumor types and stratification status. It must contain columns: SampleID, TumorType, Stratification.}

\item{df_recur_actionable}{A data frame of samples having recurrent and actionable features. It must contain columns: SampleID, Feature, Feature_Value, Feature_Recur_Pval. It can be generated by function get_recur_actionable_features.}
}
\value{
A data frame of samples that have both stratification status and recurrent/actionable features for each sample. It contains columns: SampleID, TumorType, Stratification, Feature, Feature_Value, Feature_Recur_Pval.
}
\description{
For each sample, given its stratification status and recurrent/actionable features identified by REFLECT, it combines them to form co-altered targets towards combination therapies.
}
\examples{
library(reflect)
mat_value <- egfr_data$mat_value
wbound <- 2.0
mat_value_clustered <- sparse_hclust(mat_value, wbound)$mat_value_clustered

df_feature <- egfr_data$df_feature
mat_recur_pval <- get_recur_pval(mat_value_clustered, df_feature)

recur_actionable <- get_recur_actionable_features(mat_value, mat_recur_pval)
df_recur_actionable <- recur_actionable$df_recur_actionable

df_sample <- egfr_data$df_sample
df_coaltered_targets <- get_coaltered_targets(df_sample, df_recur_actionable)
}
