% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_recur_actionable_features}
\alias{get_recur_actionable_features}
\title{Get the Recurrent and Actionable Features for Each Sample}
\usage{
get_recur_actionable_features(
  mat_value,
  mat_recur_pval,
  pval_threshold = 0.05,
  df_feature = NULL
)
}
\arguments{
\item{mat_value}{A matrix of expression/alteration with samples as rows and features as columns.}

\item{mat_recur_pval}{A matrix that contains recurrence P values. It may not have the same row/column orders as mat_value, but the set of samples/features must be same between the two matrices.}

\item{pval_threshold}{The threshold of P value below which are considered statistically significant. Default 0.05.}

\item{df_feature}{A data frame that annotates actionability of features. It must contains columns: Feature, IsActionable. The type of IsActionable is logical. If df_feature = NULL, all features are considered to be actionable.}
}
\value{
\item{df_recur_actionable}{A data frame of samples having recurrent and actionable features. It contains columns: SampleID, Feature, Feature_Value, Feature_Recur_Pval.} \item{mat_value_recur_actionable}{A matrix that only has recurrent and actionable element as in the input (mat_value).}
}
\description{
Given a matrix of expression/alteration, a matrix of recurrence P values, a P value threshold, and optionally a set of actionable features, filter the recurrent and actionable features for each sample from the cohort.
}
\examples{
library(reflect)
mat_value <- egfr_data$mat_value
wbound <- 2.0
mat_value_clustered <- sparse_hclust(mat_value, wbound)$mat_value_clustered

df_feature <- egfr_data$df_feature
mat_recur_pval <- get_recur_pval(mat_value_clustered, df_feature)

recur_actionable <- get_recur_actionable_features(mat_value, mat_recur_pval)
}
