% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_recur_pval}
\alias{get_recur_pval}
\title{Get the Recurrence P Values}
\usage{
get_recur_pval(mat_value_clustered, df_feature = NULL)
}
\arguments{
\item{mat_value_clustered}{A matrix of expression/alteration clustered by the sparse hierarchical clustering.}

\item{df_feature}{A data frame that annotates function scores of features. It must contains columns: Feature, FunctionScore. Function score is either 1 (activating) or -1 (inhibiting). If df_feature = NULL, both 1 (activating) and -1 (inhibiting) are considered for each feature.}
}
\value{
A matrix that contains recurrence P values for each element of the input (mat_value_clustered).
}
\description{
This employs a permutation approach to the recurrence P values. See References for details.
}
\examples{
library(reflect)
mat_value <- egfr_data$mat_value
wbound <- 2.0
mat_value_clustered <- sparse_hclust(mat_value, wbound)$mat_value_clustered

df_feature <- egfr_data$df_feature
mat_recur_pval <- get_recur_pval(mat_value_clustered, df_feature)
}
\references{
Li X., et al. (2020) Precision combination therapies from recurrent oncogenic co-alterations \doi{https://doi.org/10.1101/2020.06.03.132514}.
}
