% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{median_center}
\alias{median_center}
\title{Put the median of a numerical vector to 0}
\usage{
median_center(x)
}
\arguments{
\item{x}{A numberical vector (at least 2 non-NA items).}
}
\value{
The zero-median form of \code{x}.
}
\description{
Given a numerical vector x, return x - median(x), where NA are removed in the calculation of median(x)
}
\examples{
x <- 1:10
median_center(x)
}
