% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflect-package.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\description{
Pipe an object forward into a function or call expression.
}
\examples{
c(5:1) \%>\% sort()

}
