% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_bar_tumortypes}
\alias{plot_bar_tumortypes}
\title{Plot a Bar Chart of Tumor Type Distribution}
\usage{
plot_bar_tumortypes(df, main_label = NULL)
}
\arguments{
\item{df}{A data frame that has columns: SampleID, TumorType.}

\item{main_label}{Main lable. Default NULL.}
}
\description{
This function takes a data frame of samples and corresponding tumor types, calculates occurrences of each tumor type, then plots the occurences as a bar chart.
}
\examples{
library(reflect)
df <- egfr_data$df_sample[, c("SampleID", "TumorType")]
plot_bar_tumortypes(df)
}
