% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_bar_tumortypes_stack_tcga_ccl}
\alias{plot_bar_tumortypes_stack_tcga_ccl}
\title{Plot a Stacked Bar Chart of Tumor Types for TCGA and Cell Line Samples}
\usage{
plot_bar_tumortypes_stack_tcga_ccl(df, main_label = NULL)
}
\arguments{
\item{df}{A data frame that has columns: SampleID, TumorType.}
}
\description{
This function takes a data frame of samples and corresponding tumor types, separate TCGA and cell line samples, calculates occurrences of each tumor type, then plots the occurences as a stacked bar chart.
}
\examples{
library(reflect)
df <- egfr_data$df_sample[, c("SampleID", "TumorType")]
plot_bar_tumortypes_stack_tcga_ccl(df)
}
