% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_bar_weights}
\alias{plot_bar_weights}
\title{Plot a Bar Chart of Feature Weights}
\usage{
plot_bar_weights(df, n_top = NULL)
}
\arguments{
\item{df}{A data frame that has columns: Feature, Weight.}

\item{n_top}{An integer number showing how many top features are displayed. If NULL, all non-zero features are displayed.}
}
\description{
This function plots weights of non-zero features.
}
\examples{
library(reflect)
df <- data.frame(Feature = names(egfr_result$shc$weight),
                 Weight = egfr_result$shc$weight)
plot_bar_weights(df)
}
