% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gapstat}
\alias{plot_gapstat}
\title{Plot a Gap Statistic Profile}
\usage{
plot_gapstat(df, plot_tuning_parameter = FALSE)
}
\arguments{
\item{df}{A data frame that has columns: Wbound, NumberFeatures, GapStat, SD.}

\item{plot_tuning_parameter}{A logical value. If TRUE, it also draws a second plot of gap statistic as a function of tunning parameters. Default FALSE.}
}
\description{
This function plots a gap statistic profile as a function of the number of non-zero features, and optionally as a function of tunning parameters.
}
\examples{
library(reflect)
df <- egfr_result$gapstat_bestwbound$df_gapstat
plot_gapstat(df)
}
