% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflect-package.R
\docType{package}
\name{reflect-package}
\alias{reflect-package}
\alias{reflect}
\title{REFLECT}
\description{
REcurrent Features LEveraged for Combination Therapies
}
\details{
REFLECT is to accelerate discovery of combination therapies tailored to multi-omic profiles of cancer patient cohorts with matching oncogenic features. REFLECT identifies co-actionable, co-occurring oncogenic alterations that are recurrent within patient cohorts using genomic, transcriptomic and phosphoproteomic data. The REFLECT predictions are testable hypotheses on responses to drug combinations in specific patient cohorts matched to co-actionable alterations. The analysis of proteomic and transcriptomic signatures in relation to the variant information in patient cohorts also enable prediction of functional rare variants. REFLECT provides a novel, robust and widely applicable precision medicine approach and resource for matching oncogenic alterations to combination therapies.
}
\references{
Li X., et al. (2020) Precision combination therapies from recurrent oncogenic co-alterations \doi{https://doi.org/10.1101/2020.06.03.132514}.
}
\keyword{package}
