% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflect-package.R
\docType{data}
\name{reflect_color}
\alias{reflect_color}
\title{Color for Categories from REFLECT}
\format{
The format is a list containing the following elements:
\itemize{
\item df_tcga_color: Colors for TCGA tumor types.
\item df_ccl_color: Colors for cell line tumor types.
\item df_category_color: Colors for categories.
}
}
\description{
This is a collection of colors for tumor types and variant categories.
}
\examples{
reflect_color

}
\references{
Li X., et al. (2020) Precision combination therapies from recurrent oncogenic co-alterations \doi{https://doi.org/10.1101/2020.06.03.132514}.
}
\keyword{datasets}
