% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sparse_hclust}
\alias{sparse_hclust}
\title{Run a Sparse Hierarchical Clustering}
\usage{
sparse_hclust(
  mat_value,
  wbound,
  dissimilarity = c("squared.distance", "absolute.value")
)
}
\arguments{
\item{mat_value}{A matrix of expression/alteration with samples as rows and features as columns.}

\item{wbound}{A real number as the wbound used in sparse hierarchial clustering.}

\item{dissimilarity}{A string for the type of dissimilarity, either "squared.distance" or "absolute.value". Default "squared.distance".}
}
\value{
\item{hc}{An object of class hclust which describes the tree produced by the clustering process. See detail in function hclust from stats.} \item{weight}{The weights of features used in sparse hierachical clustering.} \item{features_nonzero}{Features that have nonzero weights.} \item{mat_value_clustered}{A matrix of values that are clustered by the sparse hierarchical clustering.}
}
\description{
Given a matrix of expression/alteration and the wbound value for L1 regularization, it performs a sparse hierarchical clustering.
}
\examples{
library(reflect)
mat_value <- egfr_data$mat_value
wbound <- 2.0
shc <- sparse_hclust(mat_value, wbound)
}
