% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{zscore}
\alias{zscore}
\title{Standardize a numerical vector (Z score transformation)}
\usage{
zscore(x)
}
\arguments{
\item{x}{A numberical vector (at least 2 non-NA items).}
}
\value{
The standardized form of \code{x}.
}
\description{
Given a numerical vector x, return (x - mean(x)) / sd(x), where NA are removed in the calculation of mean(x) and sd(x).
}
\examples{
x <- 1:10
zscore(x)
}
